🎨 Style: Todo Hook 분리
