+++ src/pages/todo.tsx
@@ -1,15 +1,9 @@
-import { useState } from "react"
 import { TodoForm, TodoHeader, TodoLayout, TodoList } from "features/todo/components"
-import type { Todo } from "features/todo/types"
+import { useTodo } from "features/todo/hooks"
 
 export default function Todo() {
-  const [todos, setTodos] = useState<Todo[]>([])
-  const handleCreateTodo = (todo: Todo) => setTodos((prevTodo) => [...prevTodo, todo])
-  const handleUpdateTodo = (todoId: number) => {
-    const updateTodoFn = (todo: Todo) => (todo.id === todoId ? { ...todo, isDone: !todo.isDone } : todo)
-    const updateTodo = todos.map(updateTodoFn)
-    setTodos(updateTodo)
-  }
+  const { todos, handleCreateTodo, handleUpdateTodo } = useTodo()
+
   return (
     <>
       <TodoLayout>
